not done yet
The previous 2 files uses the pybluez library which allows us to import the module bluetooth for use.
But for the previous meet up, can't import even after downloading the required libraries.
So for this module I used the socket library that is supposed to be tgt with python3. I uses the AF_BLUETOOTH family when initialising the socket, but still can't get it to run. AF_BLUETOOTH is supposed to be tgt with as long as your python version is after 3.
Do any of your know what the problem is?